feat(test): port remaining 21 Yoga unit test suites; fix five engine defects (task 145)
123 new tests (suite: 1333 -> 1456, all passing). The new coverage
exposed five mistranslations, all fixed against the C++:

1. PixelGrid half-way rounding compared the fractional part as float
   instead of double, collapsing the 1e-4 tolerance boundary
   (-3.5001 rounded to -3 instead of -4).
2. ZeroOutLayoutRecursively now resets the entire layout (positions,
   margins, cached measurements) like C++ getLayout() = {}, clones
   children before recursing, and its display:none call site is gated
   on performLayout to avoid mutating during measure-only passes.
3. Node.Clone() clears Owner (C++ YGNodeClone semantics).
4. CloneChildrenIfNeeded recursively clones the children of a cloned
   display:contents child.
5. CleanupContentsNodesRecursively rewritten to C++ parity:
   didPerformLayout parameter, contents-clone calls, full layout
   reset, and no recursion into regular children.

Recorded omissions: 16 auto-min-size extension tests (feature not in
the engine), 3 free-semantics tests (no GC equivalent), 2 misc.
Three stale hand-written tests codifying old behavior updated.

+++ source/timewarp-flexbox/Algorithm/PixelGrid.cs
@@ -51,12 +51,12 @@ internal static class PixelGrid
       ++fractional;
     }
 
-    if (Comparison.InexactEquals((float)fractional, 0))
+    if (Comparison.InexactEquals(fractional, 0.0))
     {
       // First we check if the value is already rounded
       scaledValue -= fractional;
     }
-    else if (Comparison.InexactEquals((float)fractional, 1.0f))
+    else if (Comparison.InexactEquals(fractional, 1.0))
     {
       scaledValue = scaledValue - fractional + 1.0;
     }
@@ -74,7 +74,7 @@ internal static class PixelGrid
       // Finally we just round the value
       scaledValue = scaledValue - fractional +
           (!double.IsNaN(fractional) &&
-           (fractional > 0.5 || Comparison.InexactEquals((float)fractional, 0.5f))
+           (fractional > 0.5 || Comparison.InexactEquals(fractional, 0.5))
               ? 1.0
               : 0.0);
     }
